Extract metrics values on a single place
